LayoutTests/platform/ios-simulator/editing/selection/move-by-word-visually-mac-expected.txt


======== Move By Word ====
Test 1, LTR:
Move right by one word
"abc def hij opq rst"[0, 3, 7, 11, 15, 19]
Move left by one word
"abc def hij opq rst"[19, 16, 12, 8, 4, 0]
Test 2, RTL:
Move left by one word
"abc def hij opq rst"[0, 16, 12, 8, 4, 19]
Move right by one word
"abc def hij opq rst"[19, 3, 7, 11, 15, 0]
Test 3, LTR:
Move right by one word
"ZZZ QQQ BBB CCC XXX"[0, 16, 12, 8, 4, 19]
Move left by one word
"ZZZ QQQ BBB CCC XXX"[19, 3, 7, 11, 15, 0]
Test 4, RTL:
Move left by one word
"ZZZ QQQ BBB CCC XXX"[0, 3, 7, 11, 15, 19]
Move right by one word
"ZZZ QQQ BBB CCC XXX"[19, 16, 12, 8, 4, 0]
Test 5, LTR:
Move right by one word
"abc def hij AAA BBB WWW opr uvw xyz"[0, 3, 7, 11, 20, 16, 23, 27, 31, 35]
Move left by one word
"abc def hij AAA BBB WWW opr uvw xyz"[35, 32, 28, 24, 15, 19, 12, 8, 4, 0]
Test 6, RTL:
Move left by one word
"abc def hij AAA BBB WWW opr uvw xyz"[0, 8, 4, 11, 15, 19, 23, 32, 28, 35]
Move right by one word
"abc def hij AAA BBB WWW opr uvw xyz"[35, 27, 31, 24, 20, 16, 12, 3, 7, 0]
Test 7, LTR:
Move right by one word
"abc def ZQB RIG uvw xyz"[0, 3, 7, 12, 15, 19, 23]
Move left by one word
"abc def ZQB RIG uvw xyz"[23, 20, 16, 11, 8, 4, 0]
Test 8, RTL:
Move left by one word
"abc def ZQB RIG uvw xyz"[0, 4, 7, 11, 15, 20, 23]
Move right by one word
"abc def ZQB RIG uvw xyz"[23, 19, 16, 12, 8, 3, 0]
Test 9, LTR:
Move right by one word
"aaa AAA bbb"[0, 3, 7, 11]
Move left by one word
"aaa AAA bbb"[11, 8, 4, 0]
Test 10, RTL:
Move left by one word
"aaa AAA bbb"[0, 3, 7, 11]
Move right by one word
"aaa AAA bbb"[11, 8, 4, 0]
Test 11, LTR:
Move right by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[0, 8, 4, 11, 15, 19, 23, 32, 28, 35]
Move left by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[35, 27, 31, 24, 20, 16, 12, 3, 7, 0]
Test 12, RTL:
Move left by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[0, 3, 7, 11, 20, 16, 23, 27, 31, 35]
Move right by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[35, 32, 28, 24, 15, 19, 12, 8, 4, 0]
Test 13, LTR:
Move right by one word
"AAA BBB aaa bbb WWW DDD"[0, 4, 7, 11, 15, 20, 23]
Move left by one word
"AAA BBB aaa bbb WWW DDD"[23, 19, 16, 12, 8, 3, 0]
Test 14, RTL:
Move left by one word
"AAA BBB aaa bbb WWW DDD"[0, 3, 7, 12, 15, 19, 23]
Move right by one word
"AAA BBB aaa bbb WWW DDD"[23, 20, 16, 11, 8, 4, 0]
Test 15, LTR:
Move right by one word
"ZQB abc RIG"[0, 3, 7, 11]
Move left by one word
"ZQB abc RIG"[11, 8, 4, 0]
Test 16, RTL:
Move left by one word
"ZQB abc RIG"[0, 3, 7, 11]
Move right by one word
"ZQB abc RIG"[11, 8, 4, 0]
Test 17, LTR:
Move right by one word
"abc def ghi jkl mn "[0, 3, 7, 11, 15, 18], "opq rst uvw xyz"[3, 7, 11, 15]
Move left by one word
"opq rst uvw xyz"[15, 12, 8, 4, 0], "abc def ghi jkl mn "[16, 12, 8, 4, 0]
Test 18, RTL:
Move left by one word
" abc def AAA AAA hij AAA AAA uvw xyz "[1, 5, 8, 12, 16, 20, 24, 28, 33, 36], "AAA kj AAA mn opq AAA AAA"[3, 6, 10, 13, 17, 21, 25]    FAIL expected: [" abc def AAA AAA hij AAA AAA uvw xyz "[ 1,  5,  8,  12,  16,  20,  24,  28,  32,  36, ]"AAA kj AAA mn opq AAA AAA"[ 3,  6,  10,  13,  17,  21,  25]
" abc def AAA AAA hij AAA AAA uvw xyz "[28, 33]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 32]
" abc def AAA AAA hij AAA AAA uvw xyz "[29, 33]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 32]
" abc def AAA AAA hij AAA AAA uvw xyz "[35, 33]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 32]
" abc def AAA AAA hij AAA AAA uvw xyz "[34, 33]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 32]
" abc def AAA AAA hij AAA AAA uvw xyz "[33, 36]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 32]
Move right by one word
"AAA kj AAA mn opq AAA AAA"[25, 22, 18, 14, 11, 7, 4, 0], " abc def AAA AAA hij AAA AAA uvw xyz "[32, 29, 25, 21, 17, 13, 9, 4, 1]    FAIL expected: ["AAA kj AAA mn opq AAA AAA"[ 25,  22,  18,  14,  11,  7,  4,  0, ]" abc def AAA AAA hij AAA AAA uvw xyz "[ 33,  29,  25,  21,  17,  13,  9,  4,  1]
"AAA kj AAA mn opq AAA AAA"[0], " abc def AAA AAA hij AAA AAA uvw xyz "[32]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 33]
<DIV>[0], " abc def AAA AAA hij AAA AAA uvw xyz "[32]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 33]
<DIV>[0], " abc def AAA AAA hij AAA AAA uvw xyz "[32]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 33]
<DIV>[0], " abc def AAA AAA hij AAA AAA uvw xyz "[32]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 33]
" abc def AAA AAA hij AAA AAA uvw xyz "[37, 32]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 33]
" abc def AAA AAA hij AAA AAA uvw xyz "[36, 32]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 33]
" abc def AAA AAA hij AAA AAA uvw xyz "[30, 32]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 33]
" abc def AAA AAA hij AAA AAA uvw xyz "[31, 32]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 33]
" abc def AAA AAA hij AAA AAA uvw xyz "[32, 29]   FAIL expected " abc def AAA AAA hij AAA AAA uvw xyz "[ 33]
Test 19, LTR:
Move right by one word
"abc def    hij opq"[0, 3, 7, 14, 18]
Move left by one word
"abc def    hij opq"[18, 15, 8, 4, 0]
Test 20, LTR:
Move right by one word
"    abc    def    hij    opq    "[4, 7, 14, 21, 28]
Move left by one word
"    abc    def    hij    opq    "[28, 22, 15, 8, 4]
Test 21, RTL:
Move left by one word
"    abc    def    hij    ABW    DSU    EJH    opq    rst    uvw    "[0, 18, 11, 21, 28, 35, 42, 60, 53, 63, 67]
Move right by one word
"    abc    def    hij    ABW    DSU    EJH    opq    rst    uvw    "[67, 49, 56, 46, 39, 32, 25, 7, 14, 4, 0]
Test 22, RTL:
Move left by one word
"    ABW    DSU    HJH    FUX    "[0, 7, 14, 21, 28, 32]
Move right by one word
"    ABW    DSU    HJH    FUX    "[32, 25, 18, 11, 4, 0]
Test 23, RTL:
Move left by one word
"abc def "[0], " rst uvw"[5, 1], "hij opq"[4], "abc def "[8, 4], " rst uvw"[8]
Move right by one word
" rst uvw"[8], "abc def "[3, 7], "hij opq"[3, 7], " rst uvw"[4], "abc def "[0]
Test 24, RTL:
Move left by one word
"ABD opq rst DSU "[0, 3, 8, 11, 15], "abc uvw AAA def lmn"[16, 12, 11, 4, 19], "ABW hij xyz FXX"[3, 8, 11, 15]    FAIL expected: ["ABD opq rst DSU "[ 0,  3,  8,  11,  15, ]"abc uvw AAA def lmn"[ 16,  12,  11,  4, ]"ABW hij xyz FXX"[ 3,  8,  11,  15]
"abc uvw AAA def lmn"[4, 19]   FAIL expected "ABW hij xyz FXX"[ 3]
"abc uvw AAA def lmn"[3, 19]   FAIL expected "ABW hij xyz FXX"[ 3]
"abc uvw AAA def lmn"[2, 19]   FAIL expected "ABW hij xyz FXX"[ 3]
"abc uvw AAA def lmn"[1, 19]   FAIL expected "ABW hij xyz FXX"[ 3]
Move right by one word
"ABW hij xyz FXX"[15, 12, 7, 4], "abc uvw AAA def lmn"[3, 7, 8, 15], "ABD opq rst DSU "[16, 12, 7, 4, 0]
